Extract spawn_harness_sync into harness_spawn module
Moves spawn_harness_sync (and the issue-watcher it contains) from
state.rs into a dedicated crates/server/src/harness_spawn.rs, leaving
state.rs as a pure struct definition with no spawning logic. Resolves
the check-state complexity violation on state.rs (score was 13,
threshold is 8).
